Renamed NavigationIconButton. Moved to navigation_components directory.
